chore(plugin): also rename plugin_host.rs to plugin_store.rs
