+++ crates/zeroclaw-plugins/src/component/v0/memory_component.rs
@@ -26,7 +26,7 @@ use super::bindings::memory::{
         ProceduralMessage as WitProceduralMessage,
     },
 };
-use super::plugin_host::{self, PluginStore};
+use super::plugin_store::{self, PluginStore};
 use crate::call_plugin;
 use crate::component::engine::ComponentEngine;
 use crate::error::PluginError;
@@ -75,7 +75,7 @@ impl ComponentMemory {
         wasmtime_wasi::p2::add_to_linker_async(&mut linker).map_err(PluginError::from)?;
         wasmtime_wasi_http::p2::add_only_http_to_linker_async(&mut linker)
             .map_err(PluginError::from)?;
-        plugin_host::add_to_linker_memory(&mut linker)?;
+        plugin_store::add_to_linker_memory(&mut linker)?;
         let host = PluginStore::with_permissions(&permissions).await?;
         let mut store = wasmtime::Store::new(engine.engine(), host);
