+++ crates/zeroclaw-plugins/src/component/v0/plugin_store.rs
